// packages/destination-actions/src/destinations/extole-destination/identifyUser/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * The unique user identifier
   */
  partner_user_id?: string
  /**
   * The anonymous user identifier
   */
  partner_anonymous_id?: string
  /**
   * User email address
   */
  email?: string
  /**
   * Traits of the event
   */
  traits?: {
    [k: string]: unknown
  }
  /**
   * Consumer device IP Address
   */
  extole_consumer_ip?: string
}